Separate view and API controllers
This commit separates the HomeController from the
APIController, making the latter responsible for
non-HTML responses.

It also removes some unused functionality, such as
the NiceHash payout percentage code.

Resolves #11.
