+++ config/routes.rb
@@ -3,11 +3,9 @@ Rails.application.routes.draw do
   get "/details" => "home#details"
   get "/financials" => "home#financials"
   get "/download" => "home#download"
-  get "/users" => "home#users"
-  get "/heartbeat" => "home#heartbeat"
-  get "/unheartbeat" => "home#unheartbeat"
-  get "/version" => "home#version"
+  get "/users" => "api#users"
+  get "/heartbeat" => "api#heartbeat"
+  get "/unheartbeat" => "api#unheartbeat"
+  get "/version" => "api#version"
   get "/release-notes" => "home#release_notes"
-  post "/nicehash-update" => "home#nicehash_update"
-  get "/payout-percentage" => "home#payout_percentage"
 end
